requirements.txt
discord.py==2.3.2
aiohttp
aiosqlite==0.19.0
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0